initial cleanup commit
