+++ .gitignore
@@ -2,3 +2,4 @@
 *#
 .ipynb_checkpoints
 .pyc
+map_orientations/__pycache__/*
